refactor: extract route middleware, schema patches, and device-reconcile modules
- agent-routes.ts: replace the per-handler `mcpAuth` + `withOrg(c, …)` preamble
  with `routes.use('*', …)` middleware (auth gate + orgContext); collapse the
  dead ENOENT branch in loadConfigDrivenProviderCatalog; inline the empty
  buildSkillsCatalogResponse wrapper; extract requireStoredPlatform /
  changePlatformRunState helpers for the GET/DELETE/start/stop platform routes.
  (requireSessionOrAdminPat stays a per-handler call — the admin-tier routes mix
  methods on shared paths, so a blanket `routes.use(path, …)` would be unsafe.)
- start-local.ts: move EMBEDDED_SCHEMA_PATCHES + EmbeddedSchemaPatch /
  MigrationSqlClient types to packages/server/src/db/embedded-schema-patches.ts
  (pure move, SQL unchanged); import them back.
- worker-api.ts: move the device-reconcile preamble (ensureDeviceConnectorWired /
  pauseStaleDeviceFeeds / reconcileDeviceCapabilities + DEVICE_WORKER_FRESH_INTERVAL
  and its exclusive imports) to packages/server/src/worker-api/device-reconcile.ts
  (verbatim); normalizeAdvertisedCapabilities stays since pollWorkerJob uses it.
- LandingPage.tsx: drop the unused LogoStrip import.
